+++ Cargo.toml
@@ -2,7 +2,6 @@
 members = [
 	"core",
 	"hexboard",
-	"hexgametile",
 	"storytell",
 ]
 resolver = "2"
