+++ core/Cargo.toml
@@ -5,7 +5,6 @@ edition = "2021"
 
 [dependencies]
 hexboard = {path = "../hexboard" }
-hexgametile = {path = "../hexgametile" }
 hex2d = "1.1.0"
 nannou = "0.18.1"
 structopt = "0.3"
