+++ core/src/hexagonaltile/factory.rs
@@ -1,5 +1,5 @@
-use hexgametile::hexagon::HexagonalTile;
-use hexgametile::terrain::Terrain;
+use crate::hexagonaltile::tile::HexagonalTile;
+use crate::terrain::Terrain;
 use hexboard::{Board, TileFactory, Hextile};
 use hex2d::Coordinate;
 use nannou::prelude::*;
