+++ core/src/lib.rs
@@ -1,10 +1,7 @@
-//! Hexgametile 
-//!
-//! Implements the hexagonal tiles and their properties to be used with hexboard.
 use nannou::color::encoding::Srgb;
 use nannou::color::rgb::Rgb;
 
-pub mod hexagon;
+pub mod hexagonaltile;
 pub mod terrain;
 
 ///Type alias for nannou color type
