--- hexgametile/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "hexgametile"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-hexboard = {path = "../hexboard"}
-nannou = "0.18.1"
-palette = { version = "0.5", features = [ "named" ]}
-hex2d = "1.1.0"
-image = "0.24.5"
