--- hexgametile/README.md
@@ -1,2 +1,0 @@
-# hexgametile
-Hexagonal tile package for small tactical RPG
